pyproject.toml
[project]
name = "snowflake-tutorial"
version = "0.1.0"
description = "Montrose Snowflake Introduction"
authors = [
    {name = "Eric Callahan"}
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10, <3.14"
dependencies = [
    "snowflake-connector-python (>=4.3.0,<5.0.0)",
    "snowflake-snowpark-python[pandas] (>=1.46.0,<2.0.0)",
    "geopandas (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
